Add trait `MethodWithRespHeader`
The trait `MethodWithRespHeader` is an alternative to `Method` and returns a tuple of String and Entity instead of just an Entity.
The returned String is the HTTP response header value associated with the Entity.
For example /api/v1/accounts/:id/followers returns array of Account and `Link` HTTP response header contains pagination controll information.